Eugene/2023 08 21/refactor the logic of importing transactions (#13)
* create csv template for bank-account pair

- add a factory to return the csv template per bank-account pair
- add tests

* complete templates and tests

* complete bank templates and tests

* refactor the csv upload and category mapping

- use template to upload csv for each bank/account
- add a new service to standardize category
- add tests

* enable batch delete by upload time

---------

+++ src/main/java/today/ihelio/jooq/tables/daos/AccountsDao.java
@@ -4,6 +4,7 @@
 package today.ihelio.jooq.tables.daos;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -132,14 +133,14 @@ public class AccountsDao extends DAOImpl<AccountsRecord, today.ihelio.jooq.table
      * Fetch records that have <code>init_balance BETWEEN lowerInclusive AND
      * upperInclusive</code>
      */
-    public List<today.ihelio.jooq.tables.pojos.Accounts> fetchRangeOfInitBalance(Long lowerInclusive, Long upperInclusive) {
+    public List<today.ihelio.jooq.tables.pojos.Accounts> fetchRangeOfInitBalance(BigDecimal lowerInclusive, BigDecimal upperInclusive) {
         return fetchRange(Accounts.ACCOUNTS.INIT_BALANCE, lowerInclusive, upperInclusive);
     }
 
     /**
      * Fetch records that have <code>init_balance IN (values)</code>
      */
-    public List<today.ihelio.jooq.tables.pojos.Accounts> fetchByInitBalance(Long... values) {
+    public List<today.ihelio.jooq.tables.pojos.Accounts> fetchByInitBalance(BigDecimal... values) {
         return fetch(Accounts.ACCOUNTS.INIT_BALANCE, values);
     }
 }
